Merge ChangeOwner{AndModuleClass}Traversers
I found while implementing scala-async that I needed the
equivalent functionality of `ChangeOwnerAndModuleClassTraverser`,
which I found squirrelled away in part of the lazy vals
implementation.

This functionality was at briefly (starting in 981424b376)
in the regular `ChangeOwnerTraverser` but was moved to a subclass
in a15969a6963 in response to code review [1].

But that review was actually about exposing a boolean
parameter "followModuleClass", rather than a criticism of
centralizing this pattern.

This commit speculates that "followModuleClass = true" is
*always* what you want.

[1] https://github.com/scala/scala/pull/1406#discussion_r1703370