Symbol substutition must consider ClassInfoType#parents
An upcoming change to uncurry, in which I plan to make substitution of
`lambda params -> apply method params` requires that I first to fix a
problem in symbol substition.

This situation can arise when the body of this definition:

    def owner1(a: A) = {
        class C extends M[a.B]
    }

is transplanted into a new owner that has a different symbol for `a`.

I speculated that value classes might also be prone to the
fact that symbol substitution neglected `ClassInfoType#parents`.

We can test change with Value Classes:  Partial Functions that are
dependent on value class param type used to fail with a spurious
overriding error, now they work correctly.