SI-7463,SI-8003 Correct wrong position for {select,apply}Dynamic calls
The new positions are range positions that directly refer to the
beginning and the end of the method calls in the sources instead of
simply point to the beginning of the expression. This allows the
scala-ide to semantically highlight select- and applyDynamic method
calls, because it has only to traverse the tree and apply the color
ranges to the given position ranges.

This also fixes the position marker of an error messages related
to a wrong Dynamic method signature.