SI-4287 Added test demonstrating hyperlinking to constructor's argument
This test was inspired by `test/files/run/t5603`, whose output (.check file)
needs to be updated by the next commit, because the positions associated to the
tree have slightly changed (this was needed to fix SI-4287).

The additional test should demonstrate that the change in positions isn't
relevant, and it doesn't affect functionality. In fact, hyperlinking to a
constructor's argument work as expected before and after the next commit.