SI-8027 REPL double tab regression
Where did double tab go?  "The shadow knows."

The regression was introduced by the last flurry
before we were left to wallow in whatever white
space remained.

Some xs put other xs under erasure.

It's clear that somebody's daughter walked into
the room and asked for a story, because, shockingly,
the case arrows don't line up.

We need a plug-in for Jenkins, or I guess Travis, to
fail the build if arrows and equals don't align,
because it clearly indicates a lapse of some kind.