SI-8054 Fix regression in TypeRef rebind with val overriding object
Regressed in 80767383fd / SI-7928

The condition added in that commit are necessary after refchecks but
poisonous before.

This still seems rather fragile: I wonder if `eliminateModuleDefs`
in `RefChecks` ought to flag the module symbol as `ARTIFACT` after
creating the accessor, so as to signal that it shouldn't be bound
anymore?