Update README, include doc/licenses in distro
We forgot to copy the licenses directory to the dists/ directory,
so that they weren't included in the binary distribution.

I did some house cleaning while I was at it. We no longer distribute
ant as far as I can see (we used to for the sbt build,
but that was long since removed).

Added license for ASM, and Jansi (just to be sure).