SI-6780 Better handling of cycles in in-scope implicit search
Implicit searches in the body of implicit members with inferred
types were leading to cycles. Before we used to resolve that
by saying there were no implicits in scope at all; now we just
skip the current context and still include the enclosing implicits.
Care is taken not to cache results under these circumstances.

This entails reworking `Context#implicitss` so that:

  - the implicit info cache only contains implicits from the current
    level. The List[List[_]] is now contructed on demand;
  - we can detect cycles by setting `implicitsCacheRunId` to -1 during
    the computation. The outer implicits when we encounter that.
  - we avoid caching when we hit a cycle or when the owner is uninitialized.